Implemented 'Beastmaster' NPC.
Players will be able to choose from a variety of beasts as a hunter to be their pest. Helps to reduce time taken to find new pet's. Not all pets will be avalile from this NPC, you may have to venture into the world in order to tame rare creatures.